LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/animation-nested-animation-expected.txt

Test

FAIL Animated container can create inner animation assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 255)"
